Refactor slice behavior out of the evaluation of the AST and into the type system.
