+++ supergsl/core/tests/test_eval.py
@@ -19,6 +19,7 @@ from supergsl.core.ast import (
     VariableDeclaration,
     SymbolReference,
     Slice,
+    SlicePosition,
     Constant,
     ListDeclaration,
     SequenceConstant,
@@ -130,46 +131,34 @@ class EvaluatePassTestCase(unittest.TestCase):
         self.assertEqual(result, 'YES!')
 
     def test_visit_slice(self):
-        """Visit Slice should visit start and end positions and create a child part."""
+        """Visit Slice should visit start and end positions and create a Slice type."""
         start = Mock()
         end = Mock()
-        parent_part = Mock(identifier='HIII')
-        expected_child_part = Mock()
-        parent_part.get_child_part_by_slice.return_value = expected_child_part
+
         self.eval_pass.visit.return_value = 'VISIT-RETURN-VAL'
 
         slice_node = Slice(start, end)
-        slice_node.get_slice_str = Mock(return_value='poop')
+        slice = self.eval_pass.visit_slice(slice_node)
 
-        new_part = self.eval_pass.visit_slice(slice_node, parent_part)
+        self.assertEqual(slice.start, 'VISIT-RETURN-VAL')
+        self.assertEqual(slice.end, 'VISIT-RETURN-VAL')
 
-        self.assertEqual(new_part, expected_child_part)
         self.eval_pass.visit.assert_has_calls([
-            call(start, parent_part),
-            call(end, parent_part)
+            call(start),
+            call(end)
         ])
 
-        parent_part.get_child_part_by_slice.assert_called_once_with(
-            'HIII[poop]',
-            'VISIT-RETURN-VAL',
-            'VISIT-RETURN-VAL'
-        )
-
     def test_visit_slice_position(self):
-        """Visit slice position should convert a slice position to a SeqPosition"""
+        """Visit slice position should convert a SlicePosition AST node to a Position"""
 
-        parent_part = Mock()
-        slice_position = Mock()
+        slice_position = SlicePosition(123, 'W', False)
 
-        convert_util_path = 'supergsl.core.eval.convert_slice_position_to_seq_position'
-        with patch(convert_util_path) as convert_patch:
-            convert_patch.return_value = 'HELLOOO'
+        result = self.eval_pass.visit_slice_position(
+            slice_position)
 
-            result = self.eval_pass.visit_slice_position(
-                slice_position, parent_part)
-
-            convert_patch.assert_called_once_with(parent_part, slice_position)
-            self.assertEqual(result, 'HELLOOO')
+        self.assertEqual(result.index, 123)
+        self.assertEqual(result.postfix, 'W')
+        self.assertEqual(result.approximate, False)
 
     def test_visit_list_declaration(self):
         """Create a collection with result of visiting each item node in declaration."""
