+++ supergsl/core/types/part.py
@@ -6,10 +6,13 @@ from supergsl.core.constants import (
     PART_SLICE_POSTFIX_END
 )
 
+
 from supergsl.core.types import SuperGSLType
 from supergsl.core.types.builtin import NucleotideSequence
 from supergsl.core.types.primer import PrimerPair
 from supergsl.core.exception import PartError
+from supergsl.core.types.slice import Position
+from .slice import Slice
 from .position import SeqPosition
 
 
@@ -99,17 +102,16 @@ class Part(NucleotideSequence):
         self._roles.extend(roles)
         self._roles = list(set(self._roles))
 
-    def get_child_part_by_slice(
-        self,
-        identifier : str,
-        start : SeqPosition,
-        end : SeqPosition
-    ) -> 'Part':
+    def slice(self, part_slice : Slice, identifier : Optional[str] = None) -> 'Part':
+        """Return a new part representing a sliced region."""
+        start_seq_pos = convert_slice_position_to_seq_position(self, part_slice.start)
+        end_seq_pos = convert_slice_position_to_seq_position(self, part_slice.end)
+
         return self.provider.get_child_part_by_slice(
             self,
             identifier,
-            start,
-            end
+            start_seq_pos,
+            end_seq_pos
         )
 
     def eval(self):
@@ -132,6 +134,36 @@ class Part(NucleotideSequence):
         """Display details about the SuperGSL object."""
         return '%s: %s' % (self.identifier, self.sequence)
 
+
 class LazyLoadedPart(SuperGSLType):
     def eval(self) -> SuperGSLType:
         raise NotImplementedError('Subclass to implement.')
+
+
+def convert_slice_position_to_seq_position(
+    parent_part : Part,
+    slice_position : Position
+) -> SeqPosition:
+    """Convert SuperGSL `ast.SlicePosition` to `types.position.SeqPosition` relative
+    to the parent part.
+
+    ast.SlicePosition has the following properties:
+        - index - The position on the part sequence.
+        - postfix - "S" or "E" - Whether the index is relative to the start
+            (S) or end (E) of the part.
+        - approximate - Boolean flag determines if the index is approximate
+            (True) or exact (False).
+    """
+    if slice_position.postfix == PART_SLICE_POSTFIX_START or slice_position.postfix is None:
+        # the index is relative to the start of the part.
+        return parent_part.start.get_relative_position(
+            slice_position.index,
+            slice_position.approximate)
+
+    elif slice_position.postfix == PART_SLICE_POSTFIX_END:
+        # the index is relative to the end of the part.
+        return parent_part.end.get_relative_position(
+            slice_position.index,
+            slice_position.approximate)
+
+    raise Exception('Unknown slice postfix: "%s"' % slice_position.postfix)
